+++ jsonpath/function_extensions/__init__.py
@@ -3,9 +3,13 @@ from .arguments import validate
 from .keys import keys
 from .length import length
 from .match import Match
+from .search import Search
+from .value import value
 
 __all__ = (
     "Match",
+    "Search",
+    "value",
     "keys",
     "length",
     "validate",