+++ jsonpath/path.py
@@ -187,7 +187,7 @@ class JSONPath:
 
     def empty(self) -> bool:
         """Return `True` if this path has no selectors."""
-        return bool(self._selectors)
+        return not bool(self._selectors)
 
 
 class CompoundJSONPath: